requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
resend==0.6.0
apscheduler==3.10.4
pydantic==2.5.2
python-crontab==3.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pytest-cov==4.1.0
openai>=1.0.0
PyJWT==2.10.1
pytz==2023.3.post1
python-jose[cryptography]==3.3.0 